requirements.txt
pytest==5.3.2
requests==2.22.0
allure-pytest==2.6.4
python-dotenv==0.10.3